Updated cache

+++ lib/builder/slides_loader.dart
@@ -1,13 +1,10 @@
 import 'dart:async';
+import 'dart:developer';
 
 import 'package:watcher/watcher.dart';
 
-import '../helpers/config.dart';
 import '../helpers/constants.dart';
-import '../helpers/extensions.dart';
-import '../models/asset_model.dart';
-import '../models/slide_model.dart';
-import '../services/assets_service.dart';
+import '../services/project_service.dart';
 import 'slide_parser.dart';
 import 'slides_pipeline.dart';
 
@@ -16,57 +13,26 @@ class SlidesLoader {
 
   static final instance = SlidesLoader._();
 
-  final _assetService = AssetService.instance;
-
   StreamSubscription<WatchEvent>? _listenerSub;
+  final _projectService = ProjectService.instance;
 
-  Future<void> _ensureFiles() async {
-    await kProjectRefs.assetsDir.ensureExists();
-    await kProjectRefs.generatedAssetsDir.ensureExists();
-    await kProjectRefs.markdownFile.ensureExists();
+  Future<void> _generate() async {
+    log('Generating slides...');
+    await _projectService.ensureExists();
 
-    try {
-      await _assetService.loadAssetsReference();
-    } on Exception {
-      await _assetService.saveAssetsReference([]);
-    }
+    final presentationRaw = await _projectService.loadMarkdown();
+    final deck = await _projectService.loadDeck();
+    final files = await _projectService.loadGeneratedFiles();
 
-    try {
-      await _assetService.loadConfigReference();
-    } on Exception {
-      await _assetService.saveConfigReference(const SDConfig.empty());
-    }
-
-    try {
-      await _assetService.loadSlidesReference();
-    } on Exception {
-      await _assetService.saveSlidesReference([]);
-    }
-  }
-
-  Future<void> generate() async {
-    await _ensureFiles();
-    try {
-      await _assetService.loadAssetsReference();
-    } on Exception {
-      await _assetService.saveAssetsReference([]);
-    }
-
-    final presentationRaw = await kProjectRefs.markdownFile.readAsString();
-
-    final config = await _assetService.loadConfigReference();
-
-    final slideParser = SlideParser(config: config);
+    final slideParser = SlideParser(config: deck.config);
 
     final slides = slideParser.run(presentationRaw);
-    final files = await _assetService.loadGeneratedAssets();
 
     final pipeline = SlidesPipeline(
       [
         // const ImageCachingTask(),
         const MermaidConverterTask(),
         const SlideThumbnailTask(),
-        const ImageCachingTask(),
       ],
     );
 
@@ -74,14 +40,17 @@ class SlidesLoader {
 
     for (var file in files) {
       if (!result.neededAssets.any((element) => element.path == file.path)) {
-        await file.delete();
+        if (await file.exists()) {
+          await file.delete();
+        }
       }
     }
 
-    await _assetService.saveConfigReference(config);
-    await _assetService.saveSlidesReference(result.slides);
-
-    await _assetService.saveAssetsReference(result.neededAssets);
+    await Future.wait([
+      _projectService.saveConfigRef(deck.config),
+      _projectService.saveSlidesRef(result.slides),
+      _projectService.saveAssetsRef(result.neededAssets),
+    ]);
   }
 
   void listen(
@@ -89,27 +58,14 @@ class SlidesLoader {
   ) {
     _listenerSub?.cancel();
 
-    final watcher = FileWatcher(kProjectRefs.markdownFile.path);
-    _listenerSub = watcher.events.listen((_) => onChange());
+    _listenerSub = _projectService.watcher.events.listen((_) => onChange());
   }
 
-  Future<
-      ({
-        Config config,
-        List<Slide> slides,
-        List<SlideAsset> assets,
-      })> loadFromStorage() async {
+  Future<DeckReferences> loadDeck() async {
     if (kCanRunProcess) {
-      await generate();
+      await _generate();
     }
-    final slides = await _assetService.loadSlidesReference();
-    final config = await _assetService.loadConfigReference();
-    final assets = await _assetService.loadAssetsReference();
 
-    return (
-      config: config,
-      slides: slides,
-      assets: assets,
-    );
+    return _projectService.loadDeck();
   }
 }
